requirements.txt
torch>=2.0.0
transformers>=4.35.0
pillow>=10.0.0
numpy>=1.24.0
tqdm>=4.65.0
